evolutions 内で記載した DELETE 文の誤りを訂正した。

+++ conf/evolutions/default/4.sql
@@ -7,4 +7,4 @@ insert into email_template (code, subject, body, registered) values (
 
 # --- !Downs
 
-delete from email_template where code = ;
+delete from email_template where code = 'createOwnerTemporaryRegistration';
